drug-discovery_agent/pyproject.toml
[tool.poetry]
name = "drug-discovery-agent"
version = "0.1.0"
description = "An agent for drug discovery tasks including compound analysis and literature research."
authors = ["Ryan Ye Min Thein - ryanymt@google.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-adk = ">=1.17.0"
mcp = "^1.0.0"
google-genai = "^1.21.1"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-cloud-aiplatform = "^1.132.0"
absl-py = "^2.3.1"
google-search-results = "^2.4.2"
pubchempy = "^1.0.5"
biopython = "^1.86"
nest-asyncio = "^1.6.0"
langchain = "^1.2.0"
langchain-core = "^1.2.2"
langchain-google-vertexai = "^3.2.0"
google-cloud-compute = "^1.40.0"
google-cloud-container = "^2.61.0"


[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
